Split out py3compat function to own file
