+++ tests.py
@@ -4,6 +4,7 @@
 import unittest
 import tempfile
 import pypandoc
+from pypandoc.py3compat import unicode_type
 import os
 import sys
 
@@ -199,12 +200,12 @@ class TestPypandoc(unittest.TestCase):
         # make sure that pandoc always returns unicode and does not mishandle it
         expected = u'üäöîôû{0}======{0}{0}'.format(os.linesep)
         written = pypandoc.convert(u'<h1>üäöîôû</h1>', 'md', format='html')
-        self.assertTrue(isinstance(written, pypandoc.unicode_type))
+        self.assertTrue(isinstance(written, unicode_type))
         self.assertEqualExceptForNewlineEnd(expected, written)
         bytes = u'<h1>üäöîôû</h1>'.encode("utf-8")
         written = pypandoc.convert(bytes, 'md', format='html')
         self.assertEqualExceptForNewlineEnd(expected, written)
-        self.assertTrue(isinstance(written, pypandoc.unicode_type))
+        self.assertTrue(isinstance(written, unicode_type))
 
         # Only use german umlauts in th next test, as iso-8859-15 covers that
         expected = u'üäö€{0}===={0}{0}'.format(os.linesep)
@@ -222,7 +223,7 @@ class TestPypandoc(unittest.TestCase):
         # with the right encoding it should work...
         written = pypandoc.convert(bytes, 'md', format='html', encoding="iso-8859-15")
         self.assertEqualExceptForNewlineEnd(expected, written)
-        self.assertTrue(isinstance(written, pypandoc.unicode_type))
+        self.assertTrue(isinstance(written, unicode_type))
 
     def test_conversion_from_non_plain_text_file(self):
         if 'CI' in os.environ:
